GUI: Code moved to new Map Screen component
This component should host the whole map showing aspect. This is
not finished but builds.
